Done add influxdb

+++ pkg/crawlers/crawler.go
@@ -39,10 +39,10 @@ type Ticker struct {
 
 type Bar struct {
 	Time   time.Time
-	Open   float32
-	High   float32
-	Low    float32
-	Close  float32
+	Open   float64
+	High   float64
+	Low    float64
+	Close  float64
 	Volume float64
 }
